atualização de mensagens de confirmação, rastreamento concluido, grafico de ganhos

+++ sistemarastreamento/wwwroot/views/CadastroDistribuidor/indexListar.js
@@ -16,27 +16,62 @@
     },
     excluir: function (id) {
 
-        var config = {
-            method: "DELETE",
-            headers: {
-                "Content-Type": "application/json; charset=utf-8"
-            },
-            credentials: 'include', //inclui cookies
-        };
+        const swalWithBootstrapButtons = Swal.mixin({
+            /*customClass: {
+                confirmButton: 'buttonSucess',
+                cancelButton: 'buttonDanger'
+            },*/
+            buttonsStyling: true
+        })
 
-        fetch("/CadastroDistribuidor/Excluir?id=" + id, config)
-            .then(function (dadosJson) {
-                var obj = dadosJson.json(); //deserializando
-                return obj;
-            })
-            .then(function (dadosObj) {
-                if (dadosObj.operacao) {
-                    document.getElementById(id).remove()
-                }
-            })
-            .catch(function () {
-                alert("Deu erro.")
-            })
+        swalWithBootstrapButtons.fire({
+            title: 'Confirmar Exclusão?',
+            text: "Você não poderá reverter isso!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonText: 'Sim, Deletar!',
+            cancelButtonText: 'Não, Cancelar!',
+            reverseButtons: true
+        })
+        .then((result) => {
+            if (result.isConfirmed) {
+
+                var config = {
+                    method: "DELETE",
+                    headers: {
+                        "Content-Type": "application/json; charset=utf-8"
+                    },
+                    credentials: 'include', //inclui cookies
+                };
+
+                fetch("/CadastroDistribuidor/Excluir?id=" + id, config)
+                    .then(function (dadosJson) {
+                        var obj = dadosJson.json(); //deserializando
+                        return obj;
+                    })
+                    .then(function (dadosObj) {
+                        if (dadosObj.operacao) {
+                            document.getElementById(id).remove()
+                        }
+                    })
+                    .catch(function () {
+                        alert("Deu erro.")
+                    })
+
+                swalWithBootstrapButtons.fire(
+                    'Deletado!',
+                    'Distribuidor Deletado.',
+                    'success'
+                )
+            }
+            else if (result.dismiss === Swal.DismissReason.cancel) {
+                swalWithBootstrapButtons.fire(
+                    'Cancelado',
+                    'Nada Alterado :)',
+                    'error'
+                )
+            }
+        })
     },
 
     alterar: function (id) {
@@ -158,7 +193,7 @@
                                         <a href="/CadastroDistribuidor/Editar?id=${dadosObj[i].id}"><i class='bx bx-edit'></i> Editar</a>
                                         <a data-fancybox data-type="iframe" data-src="/CadastroDistribuidor/IndexEmail?emaildist=${dadosObj[i].email}" href="javascript:;"><i class='bx bx-send'></i> Enviar E-mail</a>
                                         <a data-fancybox data-type="iframe" data-src="/CadastroDistribuidor/IndexSMS?telldist=${dadosObj[i].telefone}" href="javascript:;"><i class='bx bx-mail-send'></i> Enviar SMS</a>
-                                        <a href="javascript:indexListar.excluir(${dadosObj[i].id})" onclick="return confirm('Confirmar Exclusão?')"><i class='bx bxs-user-x'></i> Excluir</a>
+                                        <a href="javascript:indexListar.excluir(${dadosObj[i].id})"><i class='bx bxs-user-x'></i> Excluir</a>
                                   </div>
                                 </div>
                             </div>
@@ -167,7 +202,7 @@
 
                     linhas += template;
                 }
-
+                document.getElementById('retorno').innerHTML = 'Retorno - ' + dadosObj.length
                 if (dadosObj.length == 0) {
 
                     linhas = `
