+++ sistemarastreamento/wwwroot/views/LancamentoIndust/indexListar.js
@@ -16,27 +16,63 @@
     },
     excluir: function (id) {
 
-        var config = {
-            method: "DELETE",
-            headers: {
-                "Content-Type": "application/json; charset=utf-8"
-            },
-            credentials: 'include', //inclui cookies
-        };
+        const swalWithBootstrapButtons = Swal.mixin({
+            /*customClass: {
+                confirmButton: 'buttonSucess',
+                cancelButton: 'buttonDanger'
+            },*/
+            buttonsStyling: true
+        })
 
-        fetch("/LancamentoIndust/Excluir?id=" + id, config)
-            .then(function (dadosJson) {
-                var obj = dadosJson.json(); //deserializando
-                return obj;
-            })
-            .then(function (dadosObj) {
-                if (dadosObj.operacao) {
-                    document.getElementById(id).remove()
-                }
-            })
-            .catch(function () {
-                alert("Deu erro.")
-            })
+        swalWithBootstrapButtons.fire({
+            title: 'Confirmar Exclusão?',
+            text: "Você não poderá reverter isso!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonText: 'Sim, Deletar!',
+            cancelButtonText: 'Não, Cancelar!',
+            reverseButtons: true
+        })
+        .then((result) => {
+            if (result.isConfirmed) {
+
+                var config = {
+                    method: "DELETE",
+                    headers: {
+                        "Content-Type": "application/json; charset=utf-8"
+                    },
+                    credentials: 'include', //inclui cookies
+                };
+
+                fetch("/LancamentoIndust/Excluir?id=" + id, config)
+                    .then(function (dadosJson) {
+                        var obj = dadosJson.json(); //deserializando
+                        return obj;
+                    })
+                    .then(function (dadosObj) {
+                        if (dadosObj.operacao) {
+                            document.getElementById(id).remove()
+                        }
+                    })
+                    .catch(function () {
+                        alert("Deu erro.")
+                    })
+
+                swalWithBootstrapButtons.fire(
+                    'Deletado!',
+                    'Nota Deletado.',
+                    'success'
+                )
+            }
+            else if (result.dismiss === Swal.DismissReason.cancel) {
+                swalWithBootstrapButtons.fire(
+                    'Cancelado',
+                    'Nada Alterado :)',
+                    'error'
+                )
+            }
+        })  
+
     },
 
     alterar: function (id) {
@@ -163,7 +199,7 @@
                                     <i class="dropbtn fa fa-fw fa-ellipsis-v"></i>
                                     <div class="dropdown-content" id="acoesoption${dadosObj[i].id}" style="display: none;">
                                         <a data-fancybox data-type="iframe" data-src="/LancamentoIndust/IndexVisualizar?id=${dadosObj[i].id}" href="javascript:;"><i class='bx bx-show-alt'></i> Visualizar</a>
-                                        <a href="javascript:indexListar.excluir(${dadosObj[i].id})" onclick="return confirm('Confirmar Exclusão?')"><i class='bx bxs-user-x'></i> Excluir</a>
+                                        <a href="javascript:indexListar.excluir(${dadosObj[i].id})"><i class='bx bxs-user-x'></i> Excluir</a>
                                   </div>
                                 </div>
                             </div>
@@ -172,7 +208,7 @@
      
                     linhas += template;
                 }
-
+                document.getElementById('retorno').innerHTML = 'Retorno - ' + dadosObj.length
                 if (dadosObj.length == 0) {
 
                     linhas = `
